Add new schedule tags component
