+++ src/components/index.ts
@@ -83,6 +83,7 @@ export { default as DeploymentScheduleCard } from './DeploymentScheduleCard.vue'
 export { default as DeploymentSchedulesFieldset } from './DeploymentSchedulesFieldset.vue'
 export { default as DeploymentScheduleMenu } from './DeploymentScheduleMenu.vue'
 export { default as DeploymentScheduleToggle } from './DeploymentScheduleToggle.vue'
+export { default as DeploymentScheduleTags } from './DeploymentScheduleTags.vue'
 export { default as DeploymentRelationships } from './DeploymentRelationships.vue'
 export { default as DeploymentsCount } from './DeploymentsCount.vue'
 export { default as DeploymentsDeleteButton } from './DeploymentsDeleteButton.vue'
